third_party/WebKit/LayoutTests/html5lib/generated/run-tests7-write-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token <
CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M54, around October 2016. See https://www.chromestatus.com/features/5716060992962560 for more details.
../resources/tests7.dat: PASS
